+++ CookBook/App/Theme.swift
@@ -15,6 +15,7 @@ enum Theme {
     static let grassColor: UIColor = #colorLiteral(red: 0.1408461332, green: 0.5968318582, blue: 0.4974730611, alpha: 1)
     static let peachColor: UIColor = #colorLiteral(red: 0.9935668111, green: 0.7551704049, blue: 0.623813808, alpha: 1)
     static let beigeColor: UIColor = #colorLiteral(red: 0.9834302068, green: 0.9883986115, blue: 0.9234476686, alpha: 1)
+    static let orangeColor: CGColor = #colorLiteral(red: 0.9924501777, green: 0.5030021667, blue: 0.3802955449, alpha: 1)
     
     // MARK: - Style
     static let buttonCornerRadius: CGFloat = 15
