+++ CookBook/Cells + Compositional Layout/PopularCollectionViewCell.swift
@@ -19,21 +19,11 @@ class PopularCollectionViewCell: UICollectionViewCell {
     
     private let backgroundTitleView: UIView = {
        let view = UIView()
-        view.backgroundColor = .brown
+        view.backgroundColor = .clear
         view.alpha = 0.6
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
-    
-//    private let nameLabel: UILabel = {
-//       let label = UILabel()
-//        label.text = "Popular Recipes"
-//        label.textAlignment = .center
-//        label.font = .systemFont(ofSize: 16)
-//        label.textColor = .black
-//        label.translatesAutoresizingMaskIntoConstraints = false
-//        return label
-//    }()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -49,8 +39,6 @@ class PopularCollectionViewCell: UICollectionViewCell {
     func setupView() {
         clipsToBounds = true
         layer.cornerRadius = 10
-        
-//        addSubview(nameLabel)
         addSubview(popularImageView)
         addSubview(backgroundTitleView)
     }
@@ -71,10 +59,6 @@ class PopularCollectionViewCell: UICollectionViewCell {
             backgroundTitleView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 0),
             backgroundTitleView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: 0),
             backgroundTitleView.heightAnchor.constraint(equalTo: heightAnchor, constant: 0.1),
-            
-//            nameLabel.centerYAnchor.constraint(equalTo: backgroundTitleView.centerYAnchor),
-//            nameLabel.leadingAnchor.constraint(equalTo: backgroundTitleView.leadingAnchor, constant: 10),
-//            nameLabel.bottomAnchor.constraint(equalTo: backgroundTitleView.bottomAnchor, constant: -10)
         ])
     }
 }
