+++ CookBook/Resources/Assets.xcassets/OnboardingScreen.imageset/Contents.json
@@ -1,15 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "Cooking.jpg",
+      "filename" : "OnboardingScreen.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
+      "filename" : "OnboardingScreen 1.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
+      "filename" : "OnboardingScreen 2.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
