+++ CookBook/Resources/Base.lproj/LaunchScreen.storyboard
@@ -4,7 +4,6 @@
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
-        <capability name="Named colors" minToolsVersion="9.0"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -17,27 +16,34 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" image="LaunchScreen" translatesAutoresizingMaskIntoConstraints="NO" id="dvc-ey-eGo">
+                                <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                            </imageView>
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Text_Logo" highlightedImage="Text_Logo" translatesAutoresizingMaskIntoConstraints="NO" id="S9m-ZD-Fe9">
                                 <rect key="frame" x="36.333333333333343" y="364.66666666666669" width="341.33333333333326" height="166.66666666666669"/>
                             </imageView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
-                        <color key="backgroundColor" name="Beige"/>
                         <constraints>
                             <constraint firstItem="S9m-ZD-Fe9" firstAttribute="centerY" secondItem="Ze5-6b-2t3" secondAttribute="centerY" id="0am-5M-c53"/>
+                            <constraint firstItem="dvc-ey-eGo" firstAttribute="trailing" secondItem="Ze5-6b-2t3" secondAttribute="trailing" id="3A2-W9-P5F"/>
+                            <constraint firstItem="dvc-ey-eGo" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" id="3Ap-O5-NO4"/>
                             <constraint firstItem="S9m-ZD-Fe9" firstAttribute="centerX" secondItem="Ze5-6b-2t3" secondAttribute="centerX" id="8WV-Wh-oAe"/>
+                            <constraint firstAttribute="bottom" secondItem="dvc-ey-eGo" secondAttribute="bottom" id="9Ml-gE-ruP"/>
+                            <constraint firstAttribute="bottom" secondItem="dvc-ey-eGo" secondAttribute="bottom" id="Iaw-cG-QfN"/>
+                            <constraint firstItem="dvc-ey-eGo" firstAttribute="top" secondItem="Ze5-6b-2t3" secondAttribute="top" id="XOj-hP-nrG"/>
+                            <constraint firstAttribute="bottom" secondItem="dvc-ey-eGo" secondAttribute="bottom" id="pP9-AO-Ncb"/>
+                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="dvc-ey-eGo" secondAttribute="trailing" id="qwF-hq-0sN"/>
                         </constraints>
                     </view>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="iYj-Kq-Ea1" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="53" y="375"/>
+            <point key="canvasLocation" x="52.173913043478265" y="375"/>
         </scene>
     </scenes>
     <resources>
+        <image name="LaunchScreen" width="125" height="270.66665649414062"/>
         <image name="Text_Logo" width="341.33334350585938" height="166.66667175292969"/>
-        <namedColor name="Beige">
-            <color red="0.98431372640000003" green="0.98823523520000001" blue="0.92941176889999999" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
-        </namedColor>
     </resources>
 </document>
